create json export function in manager
